+++ src/stubServer.ts
@@ -129,7 +129,7 @@ async function processStubRequest(
     }
 
     if (fileContent !== undefined) {
-      const match = /_(\d+)_[a-zA-Z]+/.exec(filename);
+      const match = /_(\d+)_[A-Za-z]+/.exec(filename);
       const httpStatus = match !== null ? Number(match![1]) : 200; // 200: default HTTP status if none specified
 
       if (httpStatus === 204 /* No Content */) {